+++ spec/integrations/pro/consumption/strategies/lrj/vp/with_in_the_middle_revoked_spec.rb
@@ -45,7 +45,7 @@ start_karafka_and_wait_until do
   if DT.key?(0)
     consumer = setup_rdkafka_consumer
     consumer.subscribe(DT.topic)
-    consumer.poll(1_000)
+    5.times { consumer.poll(1_000) }
     consumer.close
 
     sleep(1)