+++ spec/integrations/pro/consumption/strategies/lrj/default/starved_revocation_on_rebalance_spec.rb
@@ -40,7 +40,7 @@ start_karafka_and_wait_until do
     if DT[:rebalanced].empty?
       consumer = setup_rdkafka_consumer
       consumer.subscribe(DT.topic)
-      consumer.poll(1_000)
+      5.times { consumer.poll(1_000) }
       consumer.close
       DT[:rebalanced] << true
     end