+++ spec/integrations/pro/consumption/strategies/lrj/default/with_revocation_synchronization_spec.rb
@@ -41,7 +41,7 @@ start_karafka_and_wait_until do
   if DT.key?(:started)
     consumer = setup_rdkafka_consumer
     consumer.subscribe(DT.topic)
-    consumer.poll(1_000)
+    5.times { consumer.poll(1_000) }
     consumer.close
 
     sleep(1)