+++ spec/integrations/production/inflight_topic_removal_spec.rb
@@ -30,7 +30,11 @@ rescue WaterDrop::Errors::ProduceError => e
   DT[:sync_errors] << e
 end
 
-assert_equal DT[:errors].last.cause.code, :unknown_partition
+# Based on exact timing librdkafka may try to deliver message but will fail, in such cases
+# the timeout will happen. In case topic removal is faster, unknown partition
+EXPECTED_ERRORS = %i[msg_timed_out unknown_partition].freeze
+
+assert EXPECTED_ERRORS.include?(DT[:errors].last.cause.code)
 
 # Sync here to force wait
 Karafka.producer.produce_sync(topic: DT.topics[1], payload: 'test1')
@@ -42,8 +46,8 @@ handler.wait(raise_response_error: false)
 
 DT[:errors].each do |error|
   if error.cause
-    assert_equal error.cause.code, :unknown_partition
+    assert EXPECTED_ERRORS.include?(error.cause.code)
   else
-    assert_equal error.code, :unknown_partition
+    assert EXPECTED_ERRORS.include?(error.code)
   end
 end