+++ lib/karafka/cli/swarm.rb
@@ -25,6 +25,7 @@ module Karafka
 
         Karafka::Server.execution_mode = :supervisor
         Karafka::Swarm::Supervisor.new.run
+        Kernel.exit(Karafka::App.config.internal.default_exit_code)
       end
     end
   end