+++ lib/karafka/cli/base.rb
@@ -112,7 +112,7 @@ module Karafka
                 *[names, option[2], option[1]].flatten
               ) { |value| options[option[0]] = value }
             end
-          end.parse(ARGV)
+          end.parse(ARGV.dup)
 
           options
         end