+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - [Refactor] Make sure all temporary topics have a `it-` prefix in their name.
 - [Refactor] Improve CI specs parallelization.
 - [Maintenance] Require `karafka-rdkafka` `>=` `0.19.1` due to usage of `#rd_kafka_global_init` and KIP-82.
+- [Maintenance] Require `karafka-web` `>=` `0.10.0`.
 - [Maintenance] Add Deimos routing patch into integration suite not to break it in the future.
 - [Maintenance] Remove Rails `7.0` specs due to upcoming EOL.
 - [Fix] Fix a case where `unknown_topic_or_part` error could leak out of the consumer on consumer shutdown.