+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - [Enhancement] Improve OSS vs. Pro specs execution isolation.
 - [Enhancement] Preload `librdkafka` code prior to forking in the Swarm mode to save memory.
 - [Enhancement] Extract errors tracker class reference into an internal `errors_tracker_class` config option (Pro).
+- [Enhancement] Add custom signal for rolling restarts with distinct exit code.
 - [Refactor] Introduce a `bin/verify_kafka_warnings` script to clean Kafka from temporary test-suite topics.
 - [Refactor] Introduce a `bin/verify_topics_naming` script to ensure proper test topics naming convention.
 - [Refactor] Make sure all temporary topics have a `it-` prefix in their name.