+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - [Enhancement] Improve OSS vs. Pro specs execution isolation.
 - [Enhancement] Preload `librdkafka` code prior to forking in the Swarm mode to save memory.
 - [Enhancement] Extract errors tracker class reference into an internal `errors_tracker_class` config option (Pro).
+- [Enhancement] Customize the multiplexing scale delay (Pro) per consumer group (Pro).
 - [Refactor] Introduce a `bin/verify_kafka_warnings` script to clean Kafka from temporary test-suite topics.
 - [Refactor] Introduce a `bin/verify_topics_naming` script to ensure proper test topics naming convention.
 - [Refactor] Make sure all temporary topics have a `it-` prefix in their name.
@@ -42,6 +43,7 @@
 - [Fix] Fix inconsistency in the logs where `Karafka::Server` originating logs would not have server id reference.
 - [Fix] Fix inconsistency in the logs where OS signal originating logs would not have server id reference.
 - [Fix] Post-fork WaterDrop instance looses some of the non-kafka settings.
+- [Fix] Max epoch tracking for early cleanup causes messages to be skipped until reload.
 
 ## 2.4.18 (2025-04-09)
 - [Fix] Make sure `Bundler.with_unbundled_env` is not called multiple times.